requirements.txt
urwid		# Console user interface library for Python
pyperclip	# Cross-platform clipboard module
PyXDG		# Cross-platform file storage (config)